WEBDEV/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
joblib==1.3.2
scikit-learn==1.3.2
google-generativeai==0.8.6
pandas==2.1.3
python-dotenv==1.0.0
